Fixed clit.

+++ libutil/util.h
@@ -4,4 +4,4 @@
 
 char* nalread(char* in, size_t* len, FILE* input);
 void* nalgrow(void* data, size_t old, size_t new);
-void cliterate(int argc, char** argv, FILE* input, void (*itf)(char*, size_t));
+void cliterate(size_t argc, char** argv, FILE* input, void (*itf)(char*, size_t));
